requirements.txt
anthropic>=0.18.0
promptlayer>=0.5.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
faker>=19.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0